+++ plone/app/querystring/registryreader.py
@@ -7,7 +7,12 @@ from zope.i18n import translate
 from zope.i18nmessageid import Message
 from zope.interface import implements
 from zope.schema.interfaces import IVocabularyFactory
+from zope.component.hooks import getSite
 import logging
+
+from Products.ZCTextIndex.interfaces import IZCTextIndex
+from Products.CMFCore.utils import getToolByName
+
 
 logger = logging.getLogger("plone.app.querystring")
 
@@ -108,9 +113,11 @@ class QuerystringRegistryReader(object):
 
     def mapSortableIndexes(self, values):
         """Map sortable indexes"""
+        catalog = getToolByName(getSite(), 'portal_catalog')._catalog
         sortables = {}
         for key, field in values.get('%s.field' % self.prefix).iteritems():
-            if field['sortable']:
+            if field['sortable'] and \
+               not IZCTextIndex.providedBy(catalog.getIndex(key)):
                 sortables[key] = values.get('%s.field.%s' % (self.prefix, key))
         values['sortable'] = sortables
         return values